requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
pandas>=1.4.0
tqdm>=4.64.0
wandb>=0.13.0
datasets>=2.0.0
transformers>=4.20.0
rasterio>=1.3.0
pillow>=9.0.0
opencv-python>=4.6.0
albumentations>=1.2.0
segmentation-models-pytorch>=0.3.0
tensorboard>=2.10.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.4.0
pytest>=7.0.0